+++ src/main/java/net/imglib2/algorithm/metrics/imagequality/ImageQualityMetrics.java
@@ -1,0 +1,8 @@
+package net.imglib2.algorithm.metrics.imagequality;
+
+public interface ImageQualityMetrics {
+
+    // SSIM
+    // PSNR
+    // MSE
+}
